requirements.txt
-e .
aws-cdk.aws-ec2==1.132.0